LayoutTests/imported/w3c/web-platform-tests/css/cssom/overflow-serialization-expected.txt


FAIL CSSOM - Overflow shorthand serialization assert_equals: Overflow-x/y longhands with same CSS-wide keyword should serialize correctly. expected "overflow: initial;" but got ""
